requirements.txt
httpx==0.15.4
pypika==0.37.1
tortoise-orm<0.16
zeep
betterproto
apns2
sanic>=20.12,<20.13
sanic-gzip
pyjwt